# staging.openapi.json
{
    "openapi":"3.0.2",
    "info": {
        "title":"Helm Api",
        "version":"1.0"
    },
    "servers": [
        {"url":"https://staging.myhelm.org"}
    ],
    "paths": {
        "/api/initialise": {
            "get": {
                "summary": "Initialise user with Helm",
                "description": "Initialise user with Helm",
                "parameters": [{
                    "in": "header",
                    "name": "X-Requested-With",
                    "required": true,
                    "example": "XMLHttpRequest",
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success response indicating current setup of user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "description": "Status of patient initialisation"
                                        }
                                    }
                                },
                                "examples": {
                                    "Login": {
                                        "value": {
                                            "status": "login"
                                        }
                                    },
                                    "SignTerms": {
                                        "value": {
                                            "status": "sign_terms"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "301": {
                        "description": "Redirect to login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "Error": {
                                            "type": "string",
                                            "description": "Login Expired"
                                        }
                                    }
                                },
                                "example": {
                                    "status": "Login Expired"
                                }
                            }
                        },
                        "headers": {
                            "Location": {
                                "schema": {
                                    "type": "string",
                                    "description": "Redirect to auth"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/initialise/terms": {
            "get": {
                "security": [{
                    "cookieAuth": []
                }],
                "parameters": [{
                    "in": "header",
                    "name": "X-Requested-With",
                    "required": true,
                    "example": "XMLHttpRequest",
                    "schema": {
                        "type": "string"
                    }
                }],
                "summary": "Get Terms",
                "description": "Get Helm terms for user to sign",
                "responses": {
                    "200": {
                        "description": "Policy resources for user to accept",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "resources": {
                                            "type": "array",
                                            "description": "Array of Policy resources to accept",
                                            "items": {
                                                "anyOf": [
                                                   { "$ref": "#/components/schemas/Policy" }
                                                ] 
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "resources": [{
                                    "resourceType": "Policy",
                                    "status": "active",
                                    "name": "Helm Privacy Notice",
                                    "narrative": "<div>Narrative</div>",
                                    "basis": "inclusive",
                                    "type": "individual",
                                    "action": "nothing",
                                    "rule": [
                                        {
                                            "access": {
                                                "organization": {
                                                    "reference": "Organization/YHCR:2b8605a6-5135-4910-85d2-0dc316a0cd54"
                                                }
                                            },
                                            "data": [
                                                {
                                                    "resource": {
                                                        "reference": "StructureDefinition/YHCR:Composition"
                                                    }
                                                }
                                            ]
                                        }
                                    ],
                                    "meta": {
                                        "tag": [
                                            {
                                                "system": "https://yhcr.nhs.uk/Source",
                                                "code": "YHCR",
                                                "display": "Yorkshire and Humber Care Record"
                                            }
                                        ],
                                        "lastUpdated": "2020-07-10T09:41:23.277Z",
                                        "versionId": 1
                                    },
                                    "id": "YHCR:1de1b887-bf37-46b0-81a7-3bdc8043a445"
                                }]}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorised",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/unauthorised"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/initialise/terms/accept": {
            "post": {
                "security": [{
                    "cookieAuth": []
                }],
                "parameters": [{
                    "in": "header",
                    "name": "X-Requested-With",
                    "required": true,
                    "example": "XMLHttpRequest",
                    "schema": {
                        "type": "string"
                    }
                }],
                "summary": "Return Policies accepted by user to generate consent",
                "requestBody": {
                    "description": "Array of Policy resources being accepted",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "anyOf": [
                                       { "$ref": "#/components/schemas/Policy" }
                                    ] 
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response indicating whether to continue login or request terms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "description": "Status of patient initialisation"
                                        }
                                    }
                                },
                                "examples": {
                                    "Login": {
                                        "value": {
                                            "status": "login"
                                        }
                                    },
                                    "SignTerms": {
                                        "value": {
                                            "status": "sign_terms"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorised",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/unauthorised"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/demographics": {
            "get": {
                "security": [{
                    "cookieAuth": []
                }],
                "parameters": [
                    {
                        "in": "header",
                        "name": "X-Requested-With",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "example": "XMLHttpRequest"
                    }
                ],
                "summary": "Get demographics of logged in user",
                "responses": {
                    "200": {
                        "description": "Patient demographics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Demographics"
                                },
                                "example": {
                                    "demographics": {
                                        "id": "9449306052",
                                        "nhsNumber": "9449306052",
                                        "gender": "male",
                                        "telephone": "0113554466",
                                        "name": "CUTHBERT SEBASTIAN JUMP",
                                        "address": "1 DANETREE CLOSE, EPSOM, SURREY, KT19 9SU",
                                        "dateOfBirth": 1177718400000,
                                        "gpName": "Not known",
                                        "gpAddress": "Not known"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorised",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/unauthorised"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/redirect": {
            "get": {
                "summary": "Get Redirect to OIDC login",
                "description": "Get login url",
                "responses": {
                    "200": {
                        "description": "Successful response of login url",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/returnUrl"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/token": {
            "get": {
                "summary": "OIDC Callback Url",
                "description": "Return Url from OIDC provider",
                "parameters": [{
                    "in": "query",
                    "name": "code",
                    "required": true,
                    "schema": {
                        "type": "string"
                    },
                    "example": "5f20d3ed-dd52-4989-934d-24cdd9a32c31",
                    "description": "Code to exchange for token"
                }],
                "responses": {
                    "302": {
                        "description": "Redirect to login",
                        "headers": {
                            "Set-Cookie": {
                                "description": "Auth cookie for site access",
                                "example": "JSESSIONID=token; Path=/;",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Location": {
                                "description": "Redirect to login",
                                "example": "/#/login",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "security": [{
                    "cookieAuth": []
                }],
                "summary": "Logout",
                "description": "Logout of system",
                "responses": {
                    "200": {
                        "description": "Successful logout, redirect back to login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/returnUrl"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Generic Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "JSESSIONID"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "Policy": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "resourceType": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "narrative": {
                        "type": "string"
                    },
                    "basis": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "action": {
                        "type": "string"
                    },
                    "rule": {
                        "type": "array",
                        "items": {

                        }
                    },
                    "meta": {
                        "type": "object"
                    }
                }
            },
            "returnUrl": {
                "properties": {
                    "returnURL": {
                        "type": "string",
                        "description": "Return url to NHS Login"
                    }
                },
                "example": {
                    "redirectURL": "https://oidc.mock.signin.nhs.uk/client?client_id=test&scope=openid&response_type=code&redirect_uri=http%3A%2F%2Fhelm-local.com%2Fauth%2Ftoken&prompt=login&nonce=test&state=test"
                }
            },
            "error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "error message"
                    }
                },
                "example": {
                    "error": "Error"
                }
            },
...
            "unauthorised": {
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Unauthorised request"
                    }
                },
                "example": {
                    "error": "Unauthorised"
                }
            },
            "Demographics": {
                "type": "object",
                "properties": {
                    "demographics": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "nhsNumber": {
                                "type": "string"
                            },
                            "gender": {
                                "type": "string"
                            },
                            "telephone": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "address": {
                                "type": "string"
                            },
                            "dateOfBirth": {
                                "type": "number"
                            },
                            "gpName": {
                                "type": "string"
                            },
                            "gpAddress": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        }
    }
}